Create MyQEntity
