+++ homeassistant/components/myq/light.py
@@ -1,19 +1,13 @@
 """Support for MyQ-Enabled lights."""
 import logging
 
-from pymyq.const import (
-    DEVICE_STATE as MYQ_DEVICE_STATE,
-    DEVICE_STATE_ONLINE as MYQ_DEVICE_STATE_ONLINE,
-    KNOWN_MODELS,
-    MANUFACTURER,
-)
 from pymyq.errors import MyQError
 
 from homeassistant.components.light import LightEntity
 from homeassistant.const import STATE_OFF, STATE_ON
 from homeassistant.exceptions import HomeAssistantError
-from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
+from . import MyQEntity
 from .const import DOMAIN, MYQ_COORDINATOR, MYQ_GATEWAY, MYQ_TO_HASS
 
 _LOGGER = logging.getLogger(__name__)
@@ -30,28 +24,14 @@ async def async_setup_entry(hass, config_entry, async_add_entities):
     )
 
 
-class MyQLight(CoordinatorEntity, LightEntity):
+class MyQLight(MyQEntity, LightEntity):
     """Representation of a MyQ light."""
 
     _attr_supported_features = 0
 
     def __init__(self, coordinator, device):
         """Initialize with API object, device id."""
-        super().__init__(coordinator)
-        self._device = device
-        self._attr_unique_id = device.device_id
-        self._attr_name = device.name
-
-    @property
-    def available(self):
-        """Return if the device is online."""
-        if not super().available:
-            return False
-
-        # Not all devices report online so assume True if its missing
-        return self._device.device_json[MYQ_DEVICE_STATE].get(
-            MYQ_DEVICE_STATE_ONLINE, True
-        )
+        super().__init__(coordinator, device)
 
     @property
     def is_on(self):
@@ -92,24 +72,3 @@ class MyQLight(CoordinatorEntity, LightEntity):
 
         # Write new state to HASS
         self.async_write_ha_state()
-
-    @property
-    def device_info(self):
-        """Return the device_info of the device."""
-        device_info = {
-            "identifiers": {(DOMAIN, self._device.device_id)},
-            "name": self._device.name,
-            "manufacturer": MANUFACTURER,
-            "sw_version": self._device.firmware_version,
-        }
-        if model := KNOWN_MODELS.get(self._device.device_id[2:4]):
-            device_info["model"] = model
-        if self._device.parent_device_id:
-            device_info["via_device"] = (DOMAIN, self._device.parent_device_id)
-        return device_info
-
-    async def async_added_to_hass(self):
-        """Subscribe to updates."""
-        self.async_on_remove(
-            self.coordinator.async_add_listener(self.async_write_ha_state)
-        )
